Elide the actual Go code in PredExpr.String and ActExpr.String.
String() is used to generate comments in the generated code.
// comments cannot contain newlines.
/* */ comments cannot contain nested /* */ comments.
The code in a predicate or action may contain either of these,
so we cannot use the code in the string representation
if we want to use the strings in comments.

This is probably for the better anyway, since it reduces comment noise. We have other means of getting the code string when needed.