Add rule templates.
A rule template is a rule with a name containing a parameter list.
A template can be _invoked_ using an identifier containing an argument list.
An invoked template is the template rule with all identifiers
in the rule expression whose names match a parameter
replaced with the identifier of the invocation argument.
```
A <- Triple<B, C, D>
B <- "b"
C <- "c"
D <- "d"
Triple<X, Y, Z> <- X Y Z
```