Re-do action types: infer the type from the return instead of needing it annotated.
This is sometimes better than the previous, and never worse in terms of noise.

Instead of:
	x y z "ast.Node":{ return ast.Node{x, y, z} }
you would type only:
	x y z { return ast.Node{x, y, z} }

Instead of
	n "int":{ return 0 }
you would type only:
	n { return 0 }

Now, instead of:
	x "string":{ return x + "xyz" }
you would type (about the same):
	x { return string(x + "xyz") }

It's not perfect. For example, in the last example, the string conversion is a bit redundant (it's just there for type inference), but in most cases it's better than the ugly "type": notation.