Make labels scope to their closest containing choice branch instead of the entire rule.
Previously labels were scoped to the rule. This meant that labels down different choice branches could never be initialized at the same time, but nonetheless needed unique names.

In practice, it ended up being common to need to do clumsy things like add number suffixes to label names to distinguish them across choice branches. This is nosiy and error prone:

	A <-
		thing0:Thing /
		"(" thing1:Thing ")" /
		"{" thing2:Thing "}"

With this change, the scope of the name is now bound to the nearest containing choice branch, so the same label name can be reused if they never will never conflict:

	A <-
		thing:Thing /
		"(" thing:Thing ")" /
		"{" thing:Thing "}"